requirements.txt
Flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
Pillow==10.2.0
socketio==0.2.1
eventlet==0.36.1